requirements.txt
asgiref==3.3.4
dj-config-url==0.1.1
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
